% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the shiny baseApp}
\usage{
launch_app(port = 3838)
}
\arguments{
\item{port}{The port, baseApp is running on (default: 3838)}
}
\value{
baseApp application
}
\description{
Launch the shiny baseApp
}
