% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launchApp}
\alias{launchApp}
\title{Launch the shiny baseApp}
\usage{
launchApp(port = 3838)
}
\arguments{
\item{port}{The port, baseApp is running on (default: 3838)}
}
\value{
baseApp application
}
\description{
Launch the shiny baseApp
}
