% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the shiny baseApp}
\usage{
launch_app(port = 3838, dir = tempdir())
}
\arguments{
\item{port}{The port, baseApp is running on (default: 3838)}

\item{dir}{A character string. The (absolute) path of ...}
}
\value{
baseApp application
}
\description{
Launch the shiny baseApp
}
\examples{
\dontrun{
launch_app()
}
}
